feat: criação da camada de modelo para as turmas, funções que recuperam materias por dia da semana
