+++ src/database.py
@@ -7,10 +7,8 @@ from dotenv import load_dotenv
 load_dotenv(os.path.join(os.getcwd(), '.env')) # carrega as variáveis do arquivo .env local ]
 @dataclass
 class DBCollections:
-    TURMAS_INGRESSANTES = "turmas_ingressantes"
+    RA_TURMAS = "ra_turmas"
     TURMAS_SALAS_HORARIOS = "turmas_salas_horarios"
-    TURMAS_AJUSTE = "turmas_ajuste"
-    TURMAS_REAJUSTE = "turmas_reajuste"
     FRETADOS = "fretados"
     CATALOGO = "catalogo"
     USUARIO =  "usuario"
