--- src/turmas/extract/turmas/files/.~lock.TURMAS_VETERANOS_2022.3.xlsx#
@@ -1,1 +1,0 @@
-,popfabio,pop-os,04.12.2022 14:46,file:///home/popfabio/.config/libreoffice/4;
