--- src/turmas/extract/turmas/system_separator.py
@@ -1,5 +1,0 @@
-import os
-
-# Determines the file separator depending on OS ('\\' for Windows and '/' for Linux)
-def sep():
-    return '\\' if os.name == 'nt' else '/'
