--- src/turmas/extract/turmas/xlsx_to_df.py
@@ -1,5 +1,0 @@
-import pandas as pd
-
-def xlsx_to_df(file,sheetname):
-    df = pd.read_excel(file,sheet_name=sheetname)
-    return df
