--- src/turmas/instrucoes_para_rodar.txt
@@ -1,1 +1,0 @@
-- Essa funcionalidade esta em refatoração, ainda não há instrucoes claras para rodar
