--- src/turmas/instrucoes_para_testar.txt
@@ -1,1 +1,0 @@
-- Ainda não há testes para essa funcionalidade
